Update scrypt N and maxmem parameters
We were not able to decrypt keys created with monetd. It turns out
that monetd uses different parameters with the scrypt library. In
particular, it uses a very high value for N ( 262144 = 10^18 ),
which requires 256MB of memory. By default, the node crypto library
doesnt use a high-enough maxmem value, so we hard-coded it to 300MB
to be on the safe side. We also updated the key creation method to
use the same N as monetd (instead of 8192 = 2^11).